refactor: 일급 콜렉션 Sections 추가
